Add files via upload
PNGs for the animation of the hands each number corresponding to the amount of medical supplies the player has. Also includes two PNG files for the text box. It was decided that the black boarder was preferred but I have included the white boarder as well just incase it blends in too much and needs more contrast once implemented in the game.